Backmerge 2.6.x to master

+++ src/SFML/Network/Ftp.cpp
@@ -555,7 +555,7 @@ Ftp::Response Ftp::DataChannel::open(Ftp::TransferMode mode)
             }
 
             // Reconstruct connection port and address
-            unsigned short port = static_cast<std::uint8_t>(data[4] * 256) + data[5];
+            unsigned short port = static_cast<std::uint16_t>(data[4] * 256 + data[5]);
             IpAddress      address(data[0], data[1], data[2], data[3]);
 
             // Connect the data channel to the server
